+++ prime-router/src/main/kotlin/serializers/Hl7Serializer.kt
@@ -391,7 +391,7 @@ class Hl7Serializer(
         return ReadResult(report, errors, warnings)
     }
 
-    internal fun createMessage(report: Report, row: Int): String {
+    fun createMessage(report: Report, row: Int): String {
         val message = ORU_R01()
         val hl7Config = report.destination?.translation as? Hl7Configuration?
         val processingId = if (hl7Config?.useTestProcessingMode == true) {
@@ -668,7 +668,7 @@ class Hl7Serializer(
     /**
      * Set the [terser]'s ORC-21 in accordance to the [useOrderingFacilityName] value.
      */
-    internal fun setOrderingFacilityComponent(
+    fun setOrderingFacilityComponent(
         terser: Terser,
         rawFacilityName: String,
         useOrderingFacilityName: Hl7Configuration.OrderingFacilityName,
@@ -730,7 +730,7 @@ class Hl7Serializer(
     /**
      * Lookup the NCES id if the site_type is a k12 school
      */
-    internal fun getSchoolId(report: Report, row: Int, rawFacilityName: String): String? {
+    fun getSchoolId(report: Report, row: Int, rawFacilityName: String): String? {
         // This code only works on the COVID-19 schema or its extensions
         if (!report.schema.containsElement("ordering_facility_name")) return null
         // This recommendation only applies to k-12 schools
