--- prime-router/src/test/resources/metadata/file_name_templates/file-name-templates.yml
@@ -1,74 +1,0 @@
----
-- name: standard
-  lowerCase: true
-  elements:
-    - schemaBaseName()
-    - "-"
-    - uuid()
-    - "-"
-    - createdDate()
-
-- name: standard_secondary
-  lowerCase: true
-  elements:
-    - schemaBaseName()
-    - "-"
-    - createdDate()
-    - "-"
-    - uuid()
-    - "-secondary"
-
-- name: ohio
-  upperCase: true
-  elements:
-    - cdcprime_
-    - createdDate()
-
-- name: aphl_light
-  lowerCase: true
-  elements:
-    - cdcprime_
-    - receivingOrganization()
-    - "_"
-    - processingModeCode()
-    - "_"
-    - createdDate()
-
-- name: aphl
-  lowerCase: true
-  elements:
-    - cdcprime_cdcprime_
-    - receivingOrganization()
-    - "_"
-    - processingModeCode()
-    - "_"
-    - processingModeCode()
-    - "_"
-    - createdDate()
-
-- name: washington
-  lowerCase: true
-  elements:
-    - cdcprime_
-    - processingModeCode()
-    - "_"
-    - createdDate()
-    - "_"
-    - rand6()
-
-- name: washington_negative
-  lowerCase: true
-  elements:
-    - cdcprime_
-    - processingModeCode()
-    - "_"
-    - createdDate()
-    - "_"
-    - rand6()
-    - "___neg"
-
-- name: massachusetts
-  lowerCase: true
-  elements:
-    - "prime-"
-    - createdDate()
