--- prime-router/src/test/resources/metadata/schemas/AZ/pima-az-covid-19-csv.schema
@@ -1,243 +1,0 @@
----
-name: pima-az-covid-19
-description: Pima County, AZ COVID-19 flat file
-topic: covid-19
-trackingElement: testing_lab_specimen_id
-basedOn: covid-19
-elements:
-  - name: patient_last_name
-    csvFields: [{name: Patient_last_name}]
-
-  - name: patient_first_name
-    csvFields: [{name: Patient_first_name}]
-
-  - name: patient_middle_name
-    csvFields: [{name: Patient_middle_name}]
-
-  - name: patient_suffix
-    csvFields: [{name: Patient_suffix}]
-
-  # Patient ID is generated internally by SimpleReport
-  - name: patient_id
-    csvFields: [{name: Patient_ID}]
-
-  # LOINC code for the test performed
-  - name: test_performed_name
-    csvFields: [{name: Ordered_test_name}]
-
-  - name: specimen_source_site_code
-    csvFields: [{name: Specimen_source_site, format: $display}]
-
-  - name: specimen_type
-    csvFields: [{name: Specimen_type_code, format: $display}]
-
-  # The type of device used for the test.  eg, BinaxNOW COVID-19 Ag Card
-  - name: equipment_model_name
-    csvFields: [{name: Device_ID}]
-
-  # TODO Badness: equipment_model_name is not a CODE type. Really should use a table.
-  # TODO: As items are added in prime-data-input-client/backend/src/main/resources/application.yaml,
-  #  we need to add them here.
-  - name: device_type
-    type: CODE
-    mapper: use(equipment_model_name)
-    documentation:   "Additional field per request from the County.  HHS lists 4 codes for 'Prior test' question that may match whats desired here:  Molecular, Antigen, Antibody/Serology, and UNK."
-    altValues:
-      - display: Antigen
-        code: "BD Veritor System for Rapid Detection of SARS-CoV-2*"
-      - display: Molecular
-        code: ID Now
-      - display: Antigen
-        code: BinaxNOW COVID-19 Ag Card
-      - display: Antigen
-        code: "LumiraDx SARS-CoV-2 Ag Test*"
-      - display: Antigen
-        code: Sofia 2 SARS Antigen FIA
-      - display: Antigen
-        code: Sofia 2 Flu + SARS Antigen FIA*
-      - display: Antigen
-        code: Sienna-Clarity COVID-19 Antigen Rapid Test Cassette*
-      - display: Molecular
-        code: Tide Labs DTPM COVID-19 RT-PCR Test*
-    csvFields:
-    - name: Device_type
-      format: $alt
-
-  # Identifier for the specific piece of equipment used.
-  - name: equipment_instance_id
-    csvFields: [{name: Instrument_ID}]
-
-  - name: testing_lab_specimen_id
-    mapper: use(placer_order_id)
-    csvFields: [{name: Testing_lab_specimen_ID}]
-
-# Removed until SimpleReport sends this column
-#  - name: test_result_status
-#    csvFields: [{name: Test_result_status, format: $display}]
-
-  - name: test_result
-    csvFields: [{name: Test_result_code, format: $display}]
-
-  - name: illness_onset_date
-    csvFields: [{name: Illness_onset_date, format: "yyyy-MM-dd"}]
-
-  - name: specimen_collection_date_time
-    csvFields: [{name: Specimen_collection_date_time, format: "yyyy-MM-dd"}]
-
-  - name: order_test_date
-    csvFields: [{name: Order_test_date, format: "yyyy-MM-dd"}]
-
-  - name: test_result_date
-    csvFields: [{name: Test_date, format: "yyyy-MM-dd"}]
-
-  - name: date_result_released
-    csvFields: [{name: Date_result_released, format: "yyyy-MM-dd"}]
-
-  - name: patient_race
-    csvFields: [{name: Patient_race, format: $display}]
-
-  - name: patient_dob
-    csvFields: [{name: Patient_DOB, format: "yyyy-MM-dd"}]
-
-  - name: patient_gender
-    csvFields: [{name: Patient_gender}]
-
-  - name: patient_ethnicity
-    csvFields: [{name: Patient_ethnicity, format: $display}]
-
-  - name: patient_street
-    csvFields: [{name: Patient_street}]
-
-  - name: patient_street2
-    csvFields: [{name: Patient_street_2}]
-
-  - name: patient_city
-    csvFields: [{name: Patient_city}]
-
-  - name: patient_state
-    csvFields: [{name: Patient_state}]
-
-  - name: patient_zip_code
-    csvFields: [{name: Patient_zip_code}]
-
-  - name: patient_phone_number
-    csvFields: [{name: Patient_phone_number}]
-
-  - name: patient_county
-    csvFields: [{name: Patient_county}]
-
-    # Eg, Staff, Resident, Visitor, Student
-  - name: patient_role
-    type: TEXT
-    csvFields: [{name: Patient_role}]
-
-  - name: employed_in_healthcare
-    csvFields: [{name: Employed_in_healthcare}]
-
-  - name: resident_congregate_setting
-    csvFields: [{name: Resident_congregate_setting}]
-
-  - name: first_test
-    csvFields: [{name: First_test}]
-
-  - name: symptomatic_for_disease
-    csvFields: [{name: Symptomatic_for_disease}]
-
-  # Nonstandard
-  - name: organization_and_lab
-    type: TEXT
-    mapper: 'concat(organization_name, testing_lab_name)'
-    csvFields: [{name: Testing_lab_name}]
-
-  - name: testing_lab_clia
-    csvFields: [{name: Testing_lab_CLIA}]
-
-  - name: testing_lab_street
-    csvFields: [{name: Testing_lab_street}]
-
-  - name: testing_lab_street2
-    csvFields: [{name: Testing_lab_street_2}]
-
-  - name: testing_lab_city
-    csvFields: [{name: Testing_lab_city}]
-
-  - name: testing_lab_state
-    csvFields: [{name: Testing_lab_state}]
-
-  - name: testing_lab_zip_code
-    csvFields: [{name: Testing_lab_zip_code}]
-
-  - name: testing_lab_phone_number
-    csvFields: [{name: Testing_lab_phone_number}]
-
-  - name: testing_lab_county
-    csvFields: [{name: Testing_lab_county}]
-
-  # Nonstandard
-  - name: organization_and_facility
-    type: TEXT
-    mapper: 'concat(organization_name, ordering_facility_name)'
-    csvFields: [{name: Ordering_facility_name}]
-
-  - name: ordering_facility_street
-    csvFields: [{name: Ordering_facility_street}]
-
-  - name: ordering_facility_street2
-    csvFields: [{name: Ordering_facility_street_2}]
-
-  - name: ordering_facility_city
-    csvFields: [{name: Ordering_facility_city}]
-
-  - name: ordering_facility_state
-    csvFields: [{name: Ordering_facility_state}]
-
-  - name: ordering_facility_zip_code
-    csvFields: [{name: Ordering_facility_zip_code}]
-
-  - name: ordering_facility_phone_number
-    csvFields: [{name: Ordering_facility_phone_number}]
-
-  - name: ordering_facility_county
-    csvFields: [{name: Ordering_facility_county}]
-
-  - name: ordering_provider_id
-    csvFields: [{name: Ordering_provider_ID}]
-
-  - name: ordering_provider_last_name
-    csvFields: [{name: Ordering_provider_last_name}]
-
-  - name: ordering_provider_first_name
-    csvFields: [{name: Ordering_provider_first_name}]
-
-  - name: ordering_provider_street
-    csvFields: [{name: Ordering_provider_street}]
-
-  - name: ordering_provider_street2
-    csvFields: [{name: Ordering_provider_street_2}]
-
-  - name: ordering_provider_city
-    csvFields: [{name: Ordering_provider_city}]
-
-  - name: ordering_provider_state
-    csvFields: [{name: Ordering_provider_state}]
-
-  - name: ordering_provider_zip_code
-    csvFields: [{name: Ordering_provider_zip_code}]
-
-  - name: ordering_provider_phone_number
-    csvFields: [{name: Ordering_provider_phone_number}]
-
-  - name: ordering_provider_county
-    csvFields: [{name: Ordering_provider_county}]
-
-  # special fields used by mappers
-  - name: organization_name
-    type: TEXT
-    default: ''
-
-  - name: ordering_facility_name
-
-  - name: testing_lab_name
-
-  - name: placer_order_id
-
