--- prime-router/src/test/resources/metadata/tables/LIVD-Supplemental-2021-06-07.csv
@@ -1,12 +1,0 @@
-Model,Test Performed LOINC Code,Testkit Name ID,Equipment UID,is_otc,is_home,is_serial,is_unproctored,fda_ref,fda_authorization
-BinaxNOW COVID-19 Antigen Self Test,94558-4,"BinaxNOW COVID-19 Antigen Self Test_Abbott Diagnostics Scarborough, Inc.",,Y,Y,Y,Y,https://www.fda.gov/media/147251/download,"This test is authorized for non-prescription home use with self collected direct anterior nasal (nares) swab samples from individuals aged 15 years or older or adult collected anterior nasal swab samples from individuals aged two years or older."
-BinaxNOW COVID-19 Ag Card 2 Home Test,94558-4,"BinaxNOW COVID-19 Ag Card 2 Home Test_Abbott Diagnostics Scarborough, Inc.",,Y,Y,Y,N,https://www.fda.gov/media/147256/download,"authorized for OTC at-home serial screening with telehealth proctor"
-BinaxNOW COVID-19 Ag 2 Card,94558-4,"BinaxNOW COVID-19 Ag 2 Card_Abbott Diagnostics Scarborough, Inc.",,Y,N,Y,N,https://www.fda.gov/media/147261/download,The BinaxNOW COVID-19 Ag 2 Card is authorized for non-prescription use by medical professionals or trained operators who are proficient in performing rapid lateral flow tests.
-BinaxNOW COVID-19 Ag Card Home Test,94558-4,10811877011337,,N,Y,N,Y,https://www.fda.gov/media/144576/download,The BinaxNOW COVID-19 Ag Card Home Test is to be performed only with the supervision of a telehealth proctor
-BD Veritor System for Rapid Detection of SARS-CoV-2*,94558-4,"BD Veritor System for Rapid Detection of SARS-CoV-2_Becton, Dickinson and Company (BD)","BD Veritor Plus System_Becton Dickinson",N,N,Y,N,https://www.fda.gov/media/139752/download,"This test is authorized for use at the Point of Care (POC), i.e., in patient care settings operating under a CLIA Certificate of Waiver, Certificate of Compliance, or Certificate of Accreditation."
-QuickVue At-Home OTC COVID-19 Test*,95209-3,QuickVue At-Home OTC COVID-19 Test_Quidel Corporation,,Y,Y,N,Y,https://www.fda.gov/media/147247/download,"This test is authorized for non-prescription home use with self-collected (unobserved) direct anterior nasal (NS) swab specimens from individuals aged 14 years and older or with adult-collected anterior NS samples from individuals aged 2 years or older."
-QuickVue At-Home COVID-19 Test*,95209-3,QuickVue At-Home COVID-19 Test_Quidel Corporation,,N,Y,N,Y,https://www.fda.gov/media/146309/download,"This test is authorized for prescription home use with self-collected (unobserved) anterior nares (NS) swab specimens directly from individuals aged 14 years and older who are suspected of COVID-19 by their healthcare provider within the first six days of the onset of symptoms."
-Ellume COVID-19 Home Test_Ellume Limited*,95209-3,Ellume COVID-19 Home Test_Ellume Limited,,Y,Y,N,Y,https://www.fda.gov/media/144457/download,"The Ellume COVID-19 Home Test is authorized for non-prescription home use"
-Cue COVID-19 Test for Home and Over The Counter (OTC) Use,95409-9,Cue COVID-19 Test for Home and Over The Counter (OTC) Use_Cue Health Inc.,,Y,Y,N,Y,https://www.fda.gov/media/146467/download,"The Cue COVID-19 Test for Home and Over The Counter (OTC) Use is authorized for non-prescription home use"
-Cue COVID-19 Test for Home and Over The Counter (OTC) Use (Proctored),95409-9,Cue COVID-19 Test for Home and Over The Counter (OTC) Use (Proctored)_Cue Health Inc.,,Y,Y,N,N,https://www.fda.gov/media/146467/download,"The Cue COVID-19 Test for Home and Over The Counter (OTC) Use is authorized for non-prescription home use"
-Lucira COVID-19 All-In-One Test Kit,96986-5,00810055970001,,N,UNK,N,UNK,https://www.fda.gov/media/143810/download,"This test is for prescription home use with self-collected nasal swab specimens in individuals aged 14 and older who are suspected of COVID-19 by their healthcare provider. This test is also authorized for use at the point of care (POC), in patient care settings operating under a CLIA Certificate of Waiver"
