+++ prime-router/src/testIntegration/kotlin/FakeReportIntegrationTests.kt
@@ -1,8 +1,8 @@
+package gov.cdc.prime.router
+
 import assertk.assertThat
 import assertk.assertions.contains
 import assertk.assertions.isNotNull
-import gov.cdc.prime.router.FakeReport
-import gov.cdc.prime.router.Metadata
 import org.junit.jupiter.api.Test
 
 class FakeReportIntegrationTests {
